refactor(domain): extract cross-module intermediate data contracts
StageOutcome / CheckResult / Violation are not checker-internal
implementation details — they are **intermediate result contracts** that
flow checker → cli → reporter (with future consumers in enforcer / audit).
They belong in a neutral home, not in one of the consumer/producer modules.

Introduces the `ac_guard.domain` layer:

  src/ac_guard/domain/__init__.py
  src/ac_guard/domain/models.py     ← home of the three dataclasses

To keep it from becoming a grab-bag, domain/__init__.py documents four
admission criteria: cross-module intermediate, pure data (dataclass, no
I/O, stdlib-only), >=2 non-test consumers, and PR-level change-impact
review. Anything that doesn't clear all four stays in its owning module.

checker/models.py becomes a re-export shim so existing imports
(`from ac_guard.checker.models import StageOutcome`) keep working.
.importlinter gains the new `domain` layer at the bottom alongside
`config | templates`.

Production code's import source is switched to ac_guard.domain.models
in checker and cli. Tests keep the old path (via the shim), zero test
changes in this commit.
