+++ .importlinter
@@ -6,16 +6,21 @@
 ; in Track B.1 of the M6 cleanup.
 ;
 ; Allowed dependencies (each module implicitly depends on itself):
-;   config, templates  — leaves (no internal deps)
+;   config, templates, domain  — leaves (no internal deps)
 ;   adapters           → config
-;   checker            → config
-;   reporter           → config
+;   checker            → config, domain
+;   reporter           → config, domain
 ;   generator          → config, adapters
 ;   enforcer           → config, checker, reporter
-;   cli                → config, generator, adapters, enforcer, checker, reporter
+;   cli                → config, generator, adapters, enforcer, checker, reporter, domain
 ;
 ; Within the same layer, modules may NOT import each other — e.g.
 ; enforcer must not import generator, checker must not import adapters.
+;
+; The `domain` layer holds cross-module *intermediate* data contracts
+; (StageOutcome / CheckResult / Violation) that flow checker → cli →
+; reporter. It is NOT a general-purpose shared/common bucket — see
+; src/ac_guard/domain/__init__.py for admission criteria.
 
 [importlinter]
 root_package = ac_guard
@@ -27,6 +32,6 @@ layers =
     cli
     enforcer | generator
     checker | reporter | adapters
-    config | templates
+    config | templates | domain
 containers =
     ac_guard
