+++ src/ac_guard/checker/__init__.py
@@ -7,7 +7,7 @@ from ac_guard.checker.core import (
     run_precommit,
     run_stage,
 )
-from ac_guard.checker.models import CheckResult, StageOutcome, Violation
+from ac_guard.domain.models import CheckResult, StageOutcome, Violation
 
 __all__ = [
     "CheckResult",
